backend/ml/requirements.txt
numpy>=1.24
pillow>=10.0
tflite-runtime>=2.14; platform_system != "Windows"
tensorflow>=2.16; platform_system == "Windows"
